+++ server/routes/index.js
@@ -14,6 +14,7 @@ import welcome from './welcome.routes';
 import area from './area.routes';
 import contentProperty from './contentProperty.routes';
 import reportProperty from './reportedProperty.routes';
+import nextPayment from './nextPayment.routes';
 
 export default (app) => {
   app.use('/api/v1/docs', swaggerUi.serve, swaggerUi.setup(swaggerSpecs, { explorer: true }));
@@ -31,4 +32,5 @@ export default (app) => {
   app.use('/api/v1/area', area);
   app.use('/api/v1/content-property', contentProperty);
   app.use('/api/v1/report-property', reportProperty);
+  app.use('/api/v1/next-payments', nextPayment);
 };
