+++ server/routes/user.routes.js
@@ -22,7 +22,6 @@ import {
   isAdmin,
   isUnverifiedVendor,
   hasValidObjectId,
-  isUserOrVendor,
 } from '../helpers/middleware';
 import UserController from '../controllers/user.controllers';
 import Upload, { UploadController } from '../helpers/uploadImage';
@@ -852,27 +851,6 @@ router.put(
   schemaValidation(unbanUserSchema),
   UserController.unbanUser,
 );
-
-/**
- * @swagger
- * path:
- *  /user/:id:
- *    get:
- *      parameters:
- *        - in: query
- *          name: token
- *          schema:
- *            type: string
- *          description: verifies user access
- *      summary: Gets a user based by its ID
- *      tags: [User]
- *      responses:
- *        '200':
- *          description: Next payments found
- *        '500':
- *          description: Internal server error
- */
-router.get('/next-payments', authenticate, isUserOrVendor, UserController.getNextPayments);
 
 /**
  * @swagger
